requirements.txt
streamlit
openai
gspread
oauth2client
google-auth
pillow